Implemented Redis

+++ src/main/kotlin/routes/userRoute.kt
@@ -1,5 +1,6 @@
 package routes
 
+import Client
 import db.*
 import connection.dbOperation
 import connection.*
@@ -29,7 +30,14 @@ fun Route.UserRoutes(dbObject:dbOperation){
                 return@post
             }
 
+            if(Client.jedis.hexists("allEmail" ,email)){
+                call.respond(HttpStatusCode.BadRequest , JsonErrorResponse("Fail" ,"$email already present"))
+                return@post
+            }
+            Client.jedis.hset("allEmail" , email , password)
             dbObject.insert(User(email, name, password))
+
+
             val response = JsonResponse("success" , User(email, name, password))
             call.respond(HttpStatusCode.Created , response)
         } catch (ex: Exception) {
@@ -43,7 +51,25 @@ fun Route.UserRoutes(dbObject:dbOperation){
             val userInput = call.receive<Credential>()
             val email = userInput.email
             val password = userInput.password
+
+
+            if(!(Client.jedis.hexists("allEmail" ,email))){
+                call.respond(HttpStatusCode.BadRequest , JsonErrorResponse("Fail" , "$email not exist"))
+                return@post
+            }
+
+            Client.jedis.hdel("allEmail" , email)
+            val setKey = "email:title:${userInput.email}"
+            val valuesToDelete = Client.jedis.smembers(setKey)
+            valuesToDelete.forEach { value ->
+                Client.jedis.del("title:$value")
+            }
+            Client.jedis.del(setKey)
+            Client.jedis.del("AllTodo:${userInput.email}")
+
+
             val b = dbObject.deleteUser(email,password)
+
             if(b) {
                 call.respond(HttpStatusCode.OK , JsonSuccess("success"))
             }
@@ -69,7 +95,9 @@ fun Route.UserRoutes(dbObject:dbOperation){
             }
             if(password != null){
                 flag = true
+                Client.jedis.hset("allEmail" , input.email , password)
                 dbObject.updatePassword(input.email,password)
+
             }
 
             if(flag) call.respond(HttpStatusCode.OK  , JsonSuccess("success"))
